Pipedata.file_length is now a bigint to cope with fastq files.
git-svn-id: file:///data/subversion/smallrna/pipeline/trunk@529 721a6f05-5756-4a90-b3fc-8e2b6c5fda0e

+++ etc/schema.sql
@@ -183,7 +183,7 @@ CREATE TABLE pipedata (
        format_type integer REFERENCES cvterm(cvterm_id) NOT NULL,
        content_type integer REFERENCES cvterm(cvterm_id) NOT NULL,
        file_name text UNIQUE NOT NULL,
-       file_length integer NOT NULL,
+       file_length bigint NOT NULL,
        generating_pipeprocess integer REFERENCES pipeprocess(pipeprocess_id)
 );
 CREATE TABLE pipeprocess_in_pipedata (
